poster_to_text/requirements.txt
flask
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
google-cloud-logging
google-cloud-dlp
db-dtypes
pandas